+++ src/index.ts
@@ -3,5 +3,5 @@ export * from './assetLoader';
 export * from './types';
 export * from './entityTemplates';
 export * from './components';
-export * from './systems';
+export * from './systems/renderingSystem';
 export type {EntityId} from 'entix-ecs';