+++ src/systems/renderingSystem.ts
@@ -1,7 +1,7 @@
 import { EntityId, EntityManager } from "entix-ecs";
-import { Engine } from "./engine";
-import { Transform, Sprite, EntityActive, Shape, Rectangle, Circle, Triangle } from "./components";
-import { Vector2 } from "./types";
+import { Engine } from "../engine";
+import { Transform, Sprite, EntityActive, Shape, Rectangle, Circle, Triangle, Text, Button } from "../components";
+import { Vector2 } from "../types";
 
 //RENDERING
 
@@ -30,7 +30,7 @@ export function renderingSystem(engine: Engine) {// some total of all, uses laye
         //Handle Shapes
         const shape = em.getComponent(id, Shape);
         if (shape) {
-            if (!shape.active) return;
+            if (!shape.active) continue;
 
             const layer: number = shape.layer;
             let drawFn: (() => void) | null = null;
@@ -50,8 +50,25 @@ export function renderingSystem(engine: Engine) {// some total of all, uses laye
                     drawFn: drawFn
                 });
             };
-
-
+        }
+        //Handle Text
+        const text = em.getComponent(id, Text);
+        if (text) {
+            if (!text.active) return;
+            drawCalls.push({
+                layer: text.layer,
+                drawFn: () => textRenderingSystem(engine, id)
+            })
+        }
+        //Handle ButtonPressArea
+        const button = em.getComponent(id, Button);
+        if (button) {
+            if (button.showPressArea) {
+                drawCalls.push({
+                    layer: button.layer,
+                    drawFn: () => buttonPressAreaRendering(engine, id)
+                })
+            }
         }
     };
 
@@ -87,7 +104,9 @@ function rectangleRenderingSystem(engine: Engine, id: EntityId) {
         const y: number = transform.globalPosition.position.y;
         const w: number = rectangle.width;
         const h: number = rectangle.height;
-        const rotationInRadians = (Math.PI * rectangle.rotation) / 180
+        const selfRotationInRadians = (Math.PI * rectangle.rotation) / 180;
+        const transformRotationInRadians = (Math.PI * transform.rotation.value) / 180
+        const totalRotationInRadians = selfRotationInRadians + transformRotationInRadians;
         const offset: Vector2 = {
             x: rectangle.centered ? -w / 2 : 0,
             y: rectangle.centered ? -h / 2 : 0
@@ -103,11 +122,17 @@ function rectangleRenderingSystem(engine: Engine, id: EntityId) {
 
         ctx.translate(x, y);
 
-        ctx.rotate(rotationInRadians);
+        ctx.rotate(totalRotationInRadians);
 
-        ctx.scale(transform.scale.value.x,transform.scale.value.y);
+        ctx.scale(transform.scale.value.x, transform.scale.value.y);
 
-        ctx.rect(offset.x, offset.y, w, h);
+        if (rectangle.rounded) {
+            ctx.roundRect(offset.x, offset.y, w, h, rectangle.roundedRadius);
+        }
+        else {
+            ctx.rect(offset.x, offset.y, w, h);
+        }
+
 
         ctx.fill();
 
@@ -121,7 +146,13 @@ function rectangleRenderingSystem(engine: Engine, id: EntityId) {
 
             ctx.beginPath();
 
-            ctx.rect(offset.x, offset.y, w, h);
+            if (rectangle.rounded) {
+                ctx.roundRect(offset.x, offset.y, w, h, rectangle.roundedRadius);
+            }
+            else {
+                ctx.rect(offset.x, offset.y, w, h);
+            }
+
 
             ctx.stroke();
 
@@ -158,9 +189,9 @@ function circleRenderingSystem(engine: Engine, id: EntityId) {
 
         ctx.beginPath();
 
-        ctx.translate(x,y);
+        ctx.translate(x, y);
 
-        ctx.scale(transform.scale.value.x,transform.scale.value.y);
+        ctx.scale(transform.scale.value.x, transform.scale.value.y);
 
         ctx.arc(0, 0, r, 0, Math.PI * 2);
 
@@ -206,7 +237,9 @@ function triangleRenderingSystem(engine: Engine, id: EntityId) {
     }
 
     const pos = transform.globalPosition.position;
-    const rotationInRadians = (Math.PI * triangle.rotation) / 180;
+    const selfRotationInRadians = (Math.PI * triangle.rotation) / 180;
+    const transformRotationInRadians = (Math.PI * transform.rotation.value) / 180
+    const totalRotationInRadians = selfRotationInRadians + transformRotationInRadians;
 
     // Step 1: calculate local triangle points relative to center or p1
     let localP1: Vector2, localP2: Vector2, localP3: Vector2;
@@ -234,8 +267,8 @@ function triangleRenderingSystem(engine: Engine, id: EntityId) {
     ctx.globalAlpha = shape.alpha;
 
     ctx.translate(pos.x, pos.y);
-    ctx.rotate(rotationInRadians);
-    ctx.scale(transform.scale.value.x,transform.scale.value.y);
+    ctx.rotate(totalRotationInRadians);
+    ctx.scale(transform.scale.value.x, transform.scale.value.y);
 
     ctx.beginPath();
     ctx.moveTo(localP1.x, localP1.y);
@@ -258,3 +291,76 @@ function triangleRenderingSystem(engine: Engine, id: EntityId) {
 
     ctx.restore();
 };
+
+//TEXT RENDERING SYSTEM, SUB SYSTEM!
+
+function textRenderingSystem(engine: Engine, id: EntityId) {
+    const em: EntityManager = engine.getEntityManager();
+    const ctx: CanvasRenderingContext2D = engine.getCtx();
+    const transform = em.getComponent(id, Transform);
+    const text = em.getComponent(id, Text);
+    if (!transform) throw new Error("TRANSFORM NOT FOUND IN TEXT RENDERING SYSTEM! " + id);
+    if (!text) throw new Error("TEXT NOT FOUND IN TEXT RENDERING SYSTEM! " + id);
+
+    const pos: Vector2 = transform.globalPosition.position;
+    const selfRotationInRadians = (Math.PI * text.rotation) / 180;
+    const transformRotationInRadians = (Math.PI * transform.rotation.value) / 180
+    const totalRotationInRadians = selfRotationInRadians + transformRotationInRadians;
+
+    ctx.save();
+
+    ctx.translate(pos.x, pos.y);
+    ctx.rotate(totalRotationInRadians);
+    ctx.scale(transform.scale.value.x, transform.scale.value.y);
+
+    //DRAWING TEXT
+
+    ctx.beginPath();
+    ctx.textAlign = 'center';
+    ctx.textBaseline = 'middle';
+    ctx.fillStyle = text.color;
+    ctx.font = `${text.size}px ${text.style}`;
+    ctx.fillText(text.content, 0, 0, 50);
+    ctx.closePath();
+
+    //OUTLINE
+
+    if (text.outlineEnabled) {
+        ctx.beginPath();
+        ctx.strokeStyle = text.outlineColor;
+        ctx.lineWidth = text.outlineWidth;
+        ctx.font = `${text.size}px ${text.style}`;
+        ctx.strokeText(text.content, 0, 0, 50);
+        ctx.closePath();
+    }
+
+    ctx.restore();
+};
+
+//BUTTON PRESSAREA RENDEING, SUB SYSTEM!
+
+function buttonPressAreaRendering(engine: Engine, id: EntityId) {
+    const em: EntityManager = engine.getEntityManager();
+    const ctx: CanvasRenderingContext2D = engine.getCtx();
+    const transform = em.getComponent(id, Transform);
+    const button = em.getComponent(id, Button);
+    if (!transform) throw new Error("TRANSFORM NOT FOUND IN BUTTON PRESSAREA RENDERING SYSTEM! " + id);
+    if (!button) throw new Error("BUTTON NOT FOUND IN BUTTON PRESSAREA RENDERING SYSTEM! " + id);
+    const pos: Vector2 = transform?.globalPosition.position;
+    const w: number = button.pressArea.x;
+    const h: number = button.pressArea.y;
+    const rotationInRadians: number = (Math.PI * transform.rotation.value) / 180;
+    if (button.showPressArea) {
+        ctx.save();
+        ctx.translate(pos.x, pos.y);
+        ctx.rotate(rotationInRadians);
+        ctx.scale(transform.scale.value.x, transform.scale.value.y);
+        ctx.beginPath();
+        ctx.globalAlpha = 0.25;
+        ctx.fillStyle = button.pressAreaShowColor;
+        ctx.rect(-w / 2, -h / 2, button.pressArea.x, button.pressArea.y);
+        ctx.fill();
+        ctx.closePath();
+        ctx.restore();
+    };
+}