+++ src/types.ts
@@ -1,5 +1,5 @@
 import { EntityId } from "entix-ecs";
-import { Circle, Rectangle, Triangle } from "./components";
+import { Button, Circle, Rectangle, Shape, Text, Triangle } from "./components";
 
 //GENERAL TYPES
 export type EngineOptions = {
@@ -11,7 +11,7 @@ export type EngineOptions = {
 export type AssetResources = {
     images_JSON_path?: string,
     audio_JSON_path?: string,
-    jsons?:Record<string, string>
+    jsons?: Record<string, string>
 }
 export type LoadedResources = {
     imagesData?: Record<string, HTMLImageElement>;
@@ -20,72 +20,113 @@ export type LoadedResources = {
 };
 //ENGINE TYPES
 export type Vector2 = {
-    x:number,
-    y:number
+    x: number,
+    y: number
 };
 export type GlobalPosition = {
-    position:Vector2//(x,y)
+    position: Vector2//(x,y)
 };
+/* //DISABLED
 export type LocalPosition = {
     position:Vector2//(x,y)
 };
+*/
 //ENTITY OPTIONS TYPES
 export type RotationOptions = {
-    value?:number
+    value?: number
 };
 export type ScaleOptions = {
-    value?:Vector2
+    value?: Vector2
 };
 export type TransformOptions = {
-    globalPosition?:Vector2,
-    localPosition?:Vector2,
-    rotation?:RotationOptions,
-    scale?:ScaleOptions
+    globalPosition?: Vector2,
+    localPosition?: Vector2,
+    rotation?: RotationOptions,
+    scale?: ScaleOptions
 };
 export type SpriteOptions = {
-    image:HTMLImageElement,
-    width:number,
-    height:number,
-    alpha?:number,
-    rotation?:number,
-    layer?:number,
-    active?:boolean
+    image: HTMLImageElement,
+    width: number,
+    height: number,
+    alpha?: number,
+    rotation?: number,
+    layer?: number,
+    active?: boolean,
+    blocksInput?: boolean
 };
 export type EntityActiveOptions = {
-    value?:boolean
+    value?: boolean
 };
 export type ParentOptions = {
-    value?:EntityId
+    value?: EntityId
 };
 export type SceneOptions = {
-    name:string,
-    onLoad?:Function
-    onUnload?:Function
+    name: string,
+    onLoad?: Function
+    onUnload?: Function
 };
 export type ShapeType = Rectangle | Circle | Triangle;
 export type ShapeOptions = {
-    shape:ShapeType,
-    color?:string,
-    outlineEnabled?:boolean,
-    outlineWidth?:number,
-    outlineColor?:string,
-    alpha?:number,
-    active?:boolean,
-    layer?:number
+    shape: ShapeType,
+    color?: string,
+    outlineEnabled?: boolean,
+    outlineWidth?: number,
+    outlineColor?: string,
+    alpha?: number,
+    active?: boolean,
+    layer?: number,
+    blocksInput?: false
 };
 export type RectangleOptions = {
-    width?:number,
-    height?:number,
-    centered?:boolean,
-    rotation?:number
+    width?: number,
+    height?: number,
+    centered?: boolean,
+    rotation?: number,
+    rounded?:boolean,
+    roundRadius?:number
 };
 export type CircleOptions = {
-    radius?:number
+    radius?: number
 };
 export type TriangleOptions = {
-    p1?:Vector2,
-    p2?:Vector2,
-    p3?:Vector2,
-    centered?:boolean,
-    rotation?:number
+    p1?: Vector2,
+    p2?: Vector2,
+    p3?: Vector2,
+    centered?: boolean,
+    rotation?: number
 };
+export type TextOptions = {
+    content?: string,
+    size?: number,
+    color?: string,
+    outlineEnabled?: boolean,
+    outlineWidth?: number,
+    outlineColor?: string,
+    alpha?: number,
+    active?: boolean,
+    layer?: number,
+    rotation?: number,
+    style?: string,
+    maxWidth?: number
+};
+export type ButtonOptions = {
+    pressArea?: Vector2
+    onJustPressed?: Function,
+    onPress?: Function,
+    onReleased?: Function,
+    onJustHovered?: Function,
+    onHovered?: Function,
+    onHoveredReleased?: Function,
+    showPressArea?: boolean,
+    pressAreaShowColor?: string,
+    layer?: number,
+    active?: boolean,
+    changeCursorToPointer?: boolean
+};
+/*
+export type RectButtonOptions = {
+    button?:Button,
+    shape?:Shape,
+    text?:Text
+};
+*/